Act on the deep-review findings for the ignore overhaul (#347)
- Partial-scope ignores no longer half-apply. The in-app toast/sound notifier,
  search, highlights, and bookmarks were still using the narrow whole-identity
  isIgnored, so a /ignore x PUBLIC (or channel/-pattern) rule hid the row but
  still toasted and leaked into those views. They now match with full event
  context (new isMessageHidden getter; notifier uses evaluate() and also
  suppresses on NOHIGHLIGHT). Nick-only sites (autocomplete, typing) keep the
  whole-identity check — they have no event to evaluate against.
- decideStamp re-gated to COUNTABLE_TYPES. from_ignored only feeds the
  countable-type queries and client hiding is independent of the stamp, so the
  old gate was correct; restoring it keeps high-churn JOIN/PART/QUIT/NICK off the
  per-event ignore path (no wasted eval + write-only data).
- Imported ignore rows route through ignoreRulesService.add instead of a raw
  INSERT, so a crafted/legacy archive can't plant an unvalidated regex, a non-ISO
  expires_at that never lapses, or a duplicate (closes the import-bypass, the
  matcher/sweeper expiry-format drift, and the import-dup gaps in one go).
- /ignore -time is clamped (~100y); an absurd value now returns a parse error
  instead of overflowing new Date().toISOString() into an uncaught RangeError.
- Matcher promoted to shared/ (shared/ignoreMatch.ts + shared/textMatch.ts);
  server/client files are thin re-export shims. Removes ~200 lines of
  hand-mirrored code and the server↔client drift the parity comment guarded.
- stripUrls only runs when a rule has a content pattern, and the URL regex is
  compiled once instead of per call.

Tests: notifier/search leak coverage via the new getter, decideStamp join
re-gate, malicious-archive import rejection, huge -time rejection. 1009 green.
