+++ server/services/importService.test.ts
@@ -324,6 +324,79 @@ describe('importFromZipBuffer — roundtrip', () => {
     expect(imported.trusted_certificates).toBe(1);
   });
 
+  it('rejects malicious ignored_masks rows on import (bad regex / non-ISO expiry)', async () => {
+    const { alice } = seedAlice();
+    const buf = await exportToBuffer(alice.id, { includeMessages: false });
+    const yauzl = await import('yauzl');
+    const { ZipArchive } = await import('archiver');
+
+    const entries = await new Promise<Map<string, Buffer>>((resolve, reject) => {
+      yauzl.fromBuffer(buf, { lazyEntries: true }, (err, zip) => {
+        if (err) return reject(err);
+        const out = new Map<string, Buffer>();
+        zip.readEntry();
+        zip.on('entry', (entry) => {
+          if (entry.fileName.endsWith('/')) return zip.readEntry();
+          zip.openReadStream(entry, (e2, stream) => {
+            if (e2) return reject(e2);
+            const chunks: Buffer[] = [];
+            stream.on('data', (c: Buffer) => chunks.push(c));
+            stream.on('end', () => {
+              out.set(entry.fileName, Buffer.concat(chunks));
+              zip.readEntry();
+            });
+            stream.on('error', reject);
+          });
+        });
+        zip.on('end', () => resolve(out));
+        zip.on('error', reject);
+      });
+    });
+
+    const data = JSON.parse(entries.get('data.json')!.toString('utf8')) as Record<
+      string,
+      Array<Record<string, unknown>>
+    >;
+    const valid = data.ignored_masks[0];
+    // A raw INSERT would have stored both of these verbatim. Clone the valid row
+    // (so the user_id/network_id FKs rekey) and corrupt the copies.
+    data.ignored_masks.push({
+      ...valid,
+      id: 999999,
+      mask: 'eviltimer',
+      pattern_kind: 'substr',
+      expires_at: 'whenever', // non-ISO → never lapses, never sweeps
+    });
+    data.ignored_masks.push({
+      ...valid,
+      id: 999998,
+      mask: 'brokenregex',
+      pattern: '(',
+      pattern_kind: 'regex', // won't compile
+      expires_at: null,
+    });
+    entries.set('data.json', Buffer.from(JSON.stringify(data)));
+
+    const archive = new ZipArchive();
+    const chunks: Buffer[] = [];
+    archive.on('data', (c: Buffer) => chunks.push(c));
+    for (const [name, content] of entries) archive.append(content, { name });
+    await archive.finalize();
+    const rebuilt = Buffer.concat(chunks);
+
+    const target = createUser(`evil_${Date.now()}_${Math.random().toString(36).slice(2, 8)}`);
+    await importFromZipBuffer(target.id, rebuilt);
+
+    const masks = (
+      db.prepare('SELECT mask FROM ignored_masks WHERE user_id = ?').all(target.id) as Array<{
+        mask: string;
+      }>
+    ).map((r) => r.mask);
+    expect(masks).toContain('spammer!*@*'); // the valid rule imported
+    expect(masks).not.toContain('brokenregex'); // invalid regex rejected
+    expect(masks).not.toContain('eviltimer'); // non-ISO expiry rejected
+  });
+
   it('imports successfully without messages section', async () => {
     const { alice } = seedAlice();
     const dave = createUser(`dave_${Date.now()}_${Math.random().toString(36).slice(2, 8)}`);
