+++ server/services/textMatch.ts
@@ -1,69 +1,6 @@
 // Copyright (c) 2026 Brad Root
 // SPDX-License-Identifier: MPL-2.0
 
-// Shared text-matching primitives used by the highlight engine and the ignore
-// content-pattern matcher. Keeping one implementation means "highlights on a
-// word" and "ignore -pattern that word" agree on word boundaries, glob
-// translation, and URL stripping. The client keeps its own mirror of these in
-// vue_client/src/utils/ignoreMatch.ts — any change here must land there too.
-
-import { createUrlRegex } from '../../shared/urlPattern.js';
-
-// 'substr' — case-(in)sensitive substring (irssi's default -pattern / stristr)
-// 'plain'  — whole-word match of a literal (word-boundary anchored)
-// 'glob'   — whole-word glob (* and ?) match
-// 'regex'  — raw regular expression
-export type TextKind = 'substr' | 'plain' | 'glob' | 'regex';
-
-export function escapeRegex(str: string): string {
-  return str.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-}
-
-export function globToRegexSource(pattern: string): string {
-  let out = '';
-  for (const ch of pattern) {
-    if (ch === '*') out += '.*';
-    else if (ch === '?') out += '.';
-    else out += escapeRegex(ch);
-  }
-  return out;
-}
-
-// Compile a text pattern into a predicate. Returns null when the pattern can't
-// compile (invalid regex) so callers can drop the rule rather than throw.
-export function buildTextTest(
-  pattern: string,
-  kind: TextKind,
-  caseSensitive: boolean,
-): ((text: string) => boolean) | null {
-  if (kind === 'substr') {
-    if (caseSensitive) return (text: string) => text.includes(pattern);
-    const needle = pattern.toLowerCase();
-    return (text: string) => text.toLowerCase().includes(needle);
-  }
-  const flags = caseSensitive ? '' : 'i';
-  let source: string;
-  if (kind === 'regex') {
-    source = pattern;
-  } else if (kind === 'glob') {
-    source = `(?:^|\\W)(?:${globToRegexSource(pattern)})(?=\\W|$)`;
-  } else {
-    // 'plain'
-    source = `(?:^|\\W)(?:${escapeRegex(pattern)})(?=\\W|$)`;
-  }
-  try {
-    const re = new RegExp(source, flags);
-    return (text: string) => re.test(text);
-  } catch {
-    return null;
-  }
-}
-
-// Blank out URLs before matching so a word inside a link — e.g. a nick that
-// happens to appear in `https://example.com/nick` — doesn't trigger a match.
-// The URL is replaced with a space (not removed) so words on either side can't
-// fuse into a false match. Uses the same URL definition the client uses to
-// auto-link, so "renders as a link" and "matched" stay in lockstep.
-export function stripUrls(text: string): string {
-  return text.replace(createUrlRegex(), ' ');
-}
+// Implementation moved to shared/ so server + client share one copy. This
+// re-export keeps the local import path (./textMatch.js) stable.
+export * from '../../shared/textMatch.js';
