+++ vue_client/src/components/SearchModal.vue
@@ -61,9 +61,10 @@ let scopedSnapshot: typeof store.$state | null = null;
 // The server already drops senders ignored when the message arrived (the
 // insert-time from_ignored stamp). This second, live pass also hides anyone
 // ignored *after* those messages were stored, and reactively restores rows on
-// /unignore without re-issuing the query.
+// /unignore without re-issuing the query. Full-context match so level/channel/
+// pattern-scoped rules are honored here too, not just whole-identity ignores.
 const visibleResults = computed(() =>
-  store.results.filter((m) => !ignores.isIgnored(m.networkId, m.nick, m.userhost ?? '')),
+  store.results.filter((m) => !ignores.isMessageHidden(m.networkId, m)),
 );
 
 const inputEl = ref<HTMLInputElement | null>(null);
